Multiple files in frontend lit tests are now using the functions `print_jaxpr` and `print_mlir`.
Instead of repeating these functions in each file, we pull them out into a new file `lit_util_printers.py`.
