+++ frontend/test/lit/lit.cfg.py
@@ -23,7 +23,7 @@ config.test_format = lit.formats.ShTest(True)
 
 # Define the file extensions to treat as test files (with the exception of this file).
 config.suffixes = [".py"]
-config.excludes = ["lit.cfg.py"]
+config.excludes = ["lit.cfg.py", "lit_util_printers.py"]
 
 # Define the root path of where to look for tests.
 config.test_source_root = path.dirname(__file__)
