+++ frontend/test/lit/test_jax_dynamic_api.py
@@ -19,30 +19,9 @@
 import jax.numpy as jnp
 import pennylane as qml
 from jax.core import ShapedArray
+from lit_util_printers import print_jaxpr, print_mlir
 
 from catalyst import qjit
-
-
-def print_attr(f, attr, *args, aot: bool = False, **kwargs):
-    """Print function attribute"""
-    name = f"TEST {f.__name__}"
-    print("\n" + "-" * len(name))
-    print(f"{name}\n")
-    res = None
-    if not aot:
-        res = f(*args, **kwargs)
-    print(getattr(f, attr))
-    return res
-
-
-def print_jaxpr(f, *args, **kwargs):
-    """Print jaxpr code of a function"""
-    return print_attr(f, "jaxpr", *args, **kwargs)
-
-
-def print_mlir(f, *args, **kwargs):
-    """Print mlir code of a function"""
-    return print_attr(f, "mlir", *args, **kwargs)
 
 
 # CHECK-LABEL: test_qjit_dynamic_argument
