+++ frontend/test/lit/test_peephole_optimizations.py
@@ -27,6 +27,7 @@ running -apply-transform-sequence.
 import shutil
 
 import pennylane as qml
+from lit_util_printers import print_jaxpr, print_mlir
 
 from catalyst import qjit
 from catalyst.debug import get_compilation_stage
@@ -43,28 +44,6 @@ def flush_peephole_opted_mlir_to_iostream(QJIT):
     """
     print(get_compilation_stage(QJIT, "QuantumCompilationPass"))
     shutil.rmtree(QJIT.__name__)
-
-
-def print_attr(f, attr, *args, aot: bool = False, **kwargs):
-    """Print function attribute"""
-    name = f"TEST {f.__name__}"
-    print("\n" + "-" * len(name))
-    print(f"{name}\n")
-    res = None
-    if not aot:
-        res = f(*args, **kwargs)
-    print(getattr(f, attr))
-    return res
-
-
-def print_jaxpr(f, *args, **kwargs):
-    """Print jaxpr code of a function"""
-    return print_attr(f, "jaxpr", *args, **kwargs)
-
-
-def print_mlir(f, *args, **kwargs):
-    """Print mlir code of a function"""
-    return print_attr(f, "mlir", *args, **kwargs)
 
 
 #
