+++ frontend/test/lit/test_split_multiple_tapes.py
@@ -25,30 +25,9 @@ from typing import Callable, Sequence
 
 import numpy as np
 import pennylane as qml
+from lit_util_printers import print_jaxpr, print_mlir
 
 from catalyst import qjit
-
-
-def print_attr(f, attr, *args, aot: bool = False, **kwargs):
-    """Print function attribute"""
-    name = f"TEST {f.__name__}"
-    print("\n" + "-" * len(name))
-    print(f"{name}\n")
-    res = None
-    if not aot:
-        res = f(*args, **kwargs)
-    print(getattr(f, attr))
-    return res
-
-
-def print_jaxpr(f, *args, **kwargs):
-    """Print jaxpr code of a function"""
-    return print_attr(f, "jaxpr", *args, **kwargs)
-
-
-def print_mlir(f, *args, **kwargs):
-    """Print mlir code of a function"""
-    return print_attr(f, "mlir", *args, **kwargs)
 
 
 def test_multiple_tape_transforms():
